# www/api-transfer/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger-Demo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get list of transfers",
                "description": "Returns list of transfers",
                "operationId": "getAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of all transers",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "notified": {
                                                    "type": "boolean"
                                                },
                                                "paying_purse": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "balance": {
                                                            "type": "number"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        },
                                                        "user": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string"
                                                                },
                                                                "document": {
                                                                    "type": "string"
                                                                },
                                                                "profile": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "receiver_purse": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "balance": {
                                                            "type": "number"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        },
                                                        "user": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string"
                                                                },
                                                                "document": {
                                                                    "type": "string"
                                                                },
                                                                "profile": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Create transfer",
                "description": "Return created new transfer",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "receiver_purse_id": {
                                        "type": "integer"
                                    },
                                    "paying_purse_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of new transfer",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "notified": {
                                                    "type": "boolean"
                                                },
                                                "paying_purse": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "balance": {
                                                            "type": "number"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        },
                                                        "user": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string"
                                                                },
                                                                "document": {
                                                                    "type": "string"
                                                                },
                                                                "profile": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "receiver_purse": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "balance": {
                                                            "type": "number"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        },
                                                        "user": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "email": {
                                                                    "type": "string"
                                                                },
                                                                "document": {
                                                                    "type": "string"
                                                                },
                                                                "profile": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "The given data was invalid",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}